refactor: move other menus to desktop folder and cleanup
