+++ packages/desktop/components/index.js
@@ -14,6 +14,7 @@ export { default as ContactCard } from './ContactCard.svelte'
 export { default as ContactMetadataTable } from './ContactMetadataTable.svelte'
 export { default as OnboardingLayout } from './OnboardingLayout.svelte'
 export { default as NetworkCard } from './NetworkCard.svelte'
+export { default as NodeListTable } from './NodeListTable.svelte'
 export { default as Proposals } from './Proposals.svelte'
 export { default as ProposalAnswer } from './ProposalAnswer.svelte'
 export { default as ProposalCard } from './ProposalCard.svelte'
