+++ packages/desktop/components/modals/index.js
@@ -1,2 +1,3 @@
 export { default as AccountSwitcherModal } from './AccountSwitcherModal.svelte'
 export { default as FilterModal } from './FilterModal.svelte'
+export { default as ProfileActionsModal } from './ProfileActionsModal.svelte'
