+++ packages/desktop/views/dashboard/Sidebar.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { DEFAULT_NETWORK_ICON, Icon as IconEnum } from '@auxiliary/icon'
     import { Indicator } from '@bloomwalletio/ui'
-    import { ProfileActionsMenu, SidebarTab } from '@components'
+    import { ProfileActionsModal, SidebarTab } from '@components'
     import { appVersionDetails } from '@core/app/stores'
     import { localize } from '@core/i18n'
     import { activeProfile } from '@core/profile/stores'
@@ -118,7 +118,7 @@
             </button>
         </div>
     </nav>
-    <ProfileActionsMenu bind:modal={profileModal} />
+    <ProfileActionsModal bind:modal={profileModal} />
 </aside>
 
 <style lang="scss">
