+++ packages/shared/src/components/buttons/index.js
@@ -1,7 +1,5 @@
 export { default as AddInputButton } from './AddInputButton.svelte'
-export { default as TokenActionsButton } from './TokenActionsButton.svelte'
 export { default as ExportStrongholdButton } from './ExportStrongholdButton.svelte'
-export { default as NodeActionsButton } from './NodeActionsButton.svelte'
 export { default as OnboardingButton } from './OnboardingButton.svelte'
 export { default as ReceiveAddressButton } from './ReceiveAddressButton.svelte'
 export { default as ReceiveButton } from './ReceiveButton.svelte'
