+++ packages/shared/src/components/modals/index.js
@@ -1,5 +1,2 @@
-export { default as TokenActionsMenuModal } from './TokenActionsMenuModal.svelte'
-export { default as CollectibleDetailsMenu } from './CollectibleDetailsMenu.svelte'
 export { default as Modal } from './Modal.svelte'
-export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
 export { default as RecipientAccountSelector } from './RecipientAccountSelector.svelte'
