+++ packages/shared/src/components/organisms/index.js
@@ -1,2 +1,1 @@
 export { default as NodeConfigurationForm } from './NodeConfigurationForm.svelte'
-export { default as NodeListTable } from './NodeListTable.svelte'
